requirements.txt
python-docx
PyPDF2
